- Add flavor to data layer - change dir name
