+++ data/authorization/proguard-rules.pro
