+++ data/authorization/src/Api/java/com/training/app/data/authorization/forgetpassword/dataprovider/ForgetPasswordDataProviderImpl.kt
@@ -1,0 +1,9 @@
+package com.training.app.data.authorization.forgetpassword.dataprovider
+
+import com.training.app.data.authorization.forgetpassword.dto.ForgetPasswordResponse
+
+class ForgetPasswordDataProviderImpl : ForgetPasswordDataProvider {
+    override suspend fun forgetPassword(email: String): ForgetPasswordResponse {
+        return ForgetPasswordResponse(isSuccess = false)
+    }
+}
