+++ data/authorization/src/main/java/com/training/app/data/authorization/forgetpassword/dataprovider/ForgetPasswordDataProvider.kt
@@ -1,0 +1,7 @@
+package com.training.app.data.authorization.forgetpassword.dataprovider
+
+import com.training.app.data.authorization.forgetpassword.dto.ForgetPasswordResponse
+
+interface ForgetPasswordDataProvider {
+    suspend fun forgetPassword(email: String): ForgetPasswordResponse
+}
