+++ data/authorization/src/main/java/com/training/app/data/authorization/forgetpassword/dto/ForgetPasswordResponse.kt
@@ -1,0 +1,3 @@
+package com.training.app.data.authorization.forgetpassword.dto
+
+data class ForgetPasswordResponse(val isSuccess: Boolean)
