+++ data/authorization/src/main/java/com/training/app/data/authorization/forgetpassword/repo/ForgetPasswordRepository.kt
@@ -1,0 +1,7 @@
+package com.training.app.data.authorization.forgetpassword.repo
+
+import com.training.app.data.authorization.forgetpassword.dto.ForgetPasswordResponse
+
+interface ForgetPasswordRepository {
+    suspend fun forgetPassword(email: String): ForgetPasswordResponse
+}
