+++ data/authorization/src/main/java/com/training/app/data/authorization/forgetpassword/repo/ForgetPasswordRepositoryImpl.kt
@@ -1,0 +1,10 @@
+package com.training.app.data.authorization.forgetpassword.repo
+
+import com.training.app.data.authorization.forgetpassword.dataprovider.ForgetPasswordDataProvider
+import com.training.app.data.authorization.forgetpassword.dto.ForgetPasswordResponse
+
+class ForgetPasswordRepositoryImpl constructor(private val dataProvider: ForgetPasswordDataProvider) : ForgetPasswordRepository {
+    override suspend fun forgetPassword(email: String): ForgetPasswordResponse {
+        return dataProvider.forgetPassword(email)
+    }
+}
