--- data_layer/authorization/src/main/java/com/training/app/trainingapp/data/authorization/forget_password/data_provider/ForgetPasswordDataProviderImpl.kt
@@ -1,9 +1,0 @@
-package com.training.app.trainingapp.data.authorization.forget_password.data_provider
-
-import com.training.app.trainingapp.data.authorization.forget_password.dto.ForgetPasswordResponse
-
-abstract class ForgetPasswordDataProviderImpl {
-
-    abstract suspend fun forgetPassword(email: String): ForgetPasswordResponse
-
-}
