--- data_layer/authorization/src/main/java/com/training/app/trainingapp/data/authorization/forget_password/data_provider/api/ForgetPasswordApiDataProvider.kt
@@ -1,10 +1,0 @@
-package com.training.app.trainingapp.data.authorization.forget_password.data_provider.api
-
-import com.training.app.trainingapp.data.authorization.forget_password.data_provider.ForgetPasswordDataProviderImpl
-import com.training.app.trainingapp.data.authorization.forget_password.dto.ForgetPasswordResponse
-
-class ForgetPasswordApiDataProvider : ForgetPasswordDataProviderImpl() {
-    override suspend fun forgetPassword(email: String): ForgetPasswordResponse {
-        return ForgetPasswordResponse(isSuccess = false)
-    }
-}
