--- data_layer/authorization/src/main/java/com/training/app/trainingapp/data/authorization/forget_password/dto/ForgetPasswordResponse.kt
@@ -1,3 +1,0 @@
-package com.training.app.trainingapp.data.authorization.forget_password.dto
-
-data class ForgetPasswordResponse(val isSuccess: Boolean)
