--- data_layer/authorization/src/main/java/com/training/app/trainingapp/data/authorization/forget_password/repo/ForgetPasswordRepository.kt
@@ -1,10 +1,0 @@
-package com.training.app.trainingapp.data.authorization.forget_password.repo
-
-import com.training.app.trainingapp.data.authorization.forget_password.data_provider.ForgetPasswordDataProviderImpl
-import com.training.app.trainingapp.data.authorization.forget_password.dto.ForgetPasswordResponse
-
-class ForgetPasswordRepository constructor(private val dataProvider: ForgetPasswordDataProviderImpl) : ForgetPasswordDataProviderImpl() {
-    override suspend fun forgetPassword(email: String): ForgetPasswordResponse {
-        return dataProvider.forgetPassword(email)
-    }
-}
